INSTA-10356: Replace ginko for the controller tests in favor of vanilla+testify tests
